opsis: Cleanup and modernize.
